Rename EnhancedImageServer to FileAndApiServer (BL-6405)
Also merge ImageServer into FileAndApiServer and delete a couple of
obsolete files.
