+++ src/BloomExe/web/controllers/ToolboxApi.cs
@@ -18,7 +18,7 @@ namespace BloomTests.web.controllers
 			this._bookSelection = _bookSelection;
 		}
 
-		public void RegisterWithServer(EnhancedImageServer server)
+		public void RegisterWithServer(FileAndApiServer server)
 		{
 			server.RegisterEndpointHandler("toolbox/enabledTools", HandleEnabledToolsRequest, true);
 			server.RegisterEndpointHandler("toolbox/fileExists", HandleFileExistsRequest, true);
