+++ src/BloomTests/Publish/ExportEpubTestsBaseClass.cs
@@ -35,7 +35,7 @@ namespace BloomTests.Publish
 		protected string _page1Data; // contents of the file "1.xhtml" (the main content of the test book, typically)
 		protected XDocument _manifestDoc; // the contents of _manifestFile as an XDocument.
 		protected XmlNamespaceManager _ns; // Set up with all the namespaces we use (See GetNamespaceManager())
-		protected static EnhancedImageServer s_testServer;
+		protected static FileAndApiServer s_testServer;
 		protected static BookSelection s_bookSelection;
 		protected BookServer _bookServer;
 		protected string _defaultSourceValue;
@@ -59,9 +59,9 @@ namespace BloomTests.Publish
 		}
 
 
-		internal static EnhancedImageServer GetTestServer()
+		internal static FileAndApiServer GetTestServer()
 		{
-			var server = new EnhancedImageServer(new RuntimeImageProcessor(new BookRenamedEvent()), null, GetTestBookSelection(), GetTestFileLocator());
+			var server = new FileAndApiServer(new RuntimeImageProcessor(new BookRenamedEvent()), null, GetTestBookSelection(), GetTestFileLocator());
 			server.StartListening();
 			return server;
 		}
