+++ src/BloomTests/web/controllers/ReadersApiTests.cs
@@ -8,13 +8,13 @@ namespace BloomTests.web
 	[TestFixture]
 	public class ReadersApiTests
 	{
-		private EnhancedImageServer _server;
+		private FileAndApiServer _server;
 		[SetUp]
 		public void Setup()
 		{
 			var bookSelection = new BookSelection();
 			bookSelection.SelectBook(new Bloom.Book.Book());
-			_server = new EnhancedImageServer(bookSelection);
+			_server = new FileAndApiServer(bookSelection);
 
 			//needed to avoid a check in the server
 			_server.CurrentCollectionSettings = new CollectionSettings();
